Haven file refactor

+++ commands/teardown/teardownUsers.js
@@ -2,12 +2,16 @@ import { iam } from "../../aws/services.js";
 import getUserAccessKey from "../../aws/iam/users/getUserAccessKey.js";
 import deleteUserAccessKey from "../../aws/iam/users/deleteUserAccessKey.js";
 import deleteUser from "../../aws/iam/users/deleteUser.js";
+import { adminUserName } from "../../utils/config.js";
 
 const teardownUser = async (username) => {
   try {
-  	// The code below assumes a user will only have one access key, so our code
-  	// should never create a second access key.
-  	// TODO: check if user-creation code could ever create a second access key.
+    if (username === adminUserName) {
+      return Promise.resolve();
+    }
+    // The code below assumes a user will only have one access key, so our code
+    // should never create a second access key.
+    // TODO: check if user-creation code could ever create a second access key.
     const accessKeyId = await getUserAccessKey(username);
     if (accessKeyId) await deleteUserAccessKey(accessKeyId, username);
